requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
APScheduler==3.10.4
attrs==23.2.0
certifi==2024.7.4
charset-normalizer==3.3.2
contourpy==1.2.1
cycler==0.12.1
fonttools==4.53.1
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
ical==8.1.1
idna==3.7
kiwisolver==1.4.5
mashumaro==3.13.1
matplotlib==3.9.0
multidict==6.0.5
numpy==2.0.1
packaging==24.1
pillow==10.4.0
pycryptodome==3.20.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.2
pyrainbird==6.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
yarl==1.9.4
